Refactored code
Code Refactored for XML reading.
