+++ Accenture.UIPAth.CodeReviewTool/Accenture.UIPAth.CodeReviewTool.csproj
@@ -61,6 +61,7 @@
     </Compile>
     <Compile Include="Infrastructure\CopyToClipboard.cs" />
     <Compile Include="Infrastructure\ExportToExcel.cs" />
+    <Compile Include="Infrastructure\NamingCase.cs" />
     <Compile Include="Infrastructure\ReadFlowchart.cs" />
     <Compile Include="Infrastructure\ReadSequence.cs" />
     <Compile Include="Infrastructure\ReadStateMachine.cs" />
